// src/app/model/Iproveedores.ts
export interface Proveedores {
idProveedor?:string;
idProveedores?:string;
    contactoProv?: string;
    telefonoProv?: string;
    id?: number | string;
    // imagenProducto?:string;
    razonSocial?: string;
    nombreFantasia?: string;
    contacto?: string;
    direccion?: string;
    codPostal?: number | string;
    cuil?: string;
    telefono?: number | string;
    localidad?: string;
    observacionesProv?: string;
    ts_update?: Date;

  }